multi_agent_system/requirements.txt
# Core LangChain and AI
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-core>=0.1.0
groq>=0.4.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
fake-useragent>=1.4.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
jsonlines>=3.1.0

# CLI and utilities
click>=8.1.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Async support
asyncio
aiohttp>=3.8.0

# Optional: Enhanced NLP
# spacy>=3.7.0
# nltk>=3.8.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
